requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2